Extract AccessibilityViewCodingService
